Move replay API from SentryAndroid to Sentry
